refactor(components): updates to jsx extension and pascal case names
