--- site/src/components/layout/header.js
@@ -1,27 +1,0 @@
-import {Link} from 'gatsby';
-import React from 'react';
-import { HeaderStyle, HeaderNav } from '../../styles/_layout/_header.style';
-
-const Header = () => {
-    
-    return (
-        <HeaderStyle>
-            <p className='content'>
-                sam
-            </p>
-            <HeaderNav
-                className='content links'
-            >
-                <li>
-                    <Link 
-                        to="mailto:shung93uk@gmail.com" 
-                    >
-                        contact
-                    </Link>
-                </li>
-            </HeaderNav>
-        </HeaderStyle>
-    )
-};
-
-export default Header;
